+++ src/setDummy.js
@@ -1,5 +1,5 @@
 function setDummy ( dummyRequests ) {
-return function setDummy (store,key,fn) {   // Argument 'fn' should return a promise
+return function setDummy ([key,store='default'], fn ) {   // Argument 'fn' should return a promise
         dummyRequests[`${store}/${key}`] = fn
 }}